squashme: fix tests
