+++ test/components/renku_data_services/notebooks/test_notebooks_image_checks.py
@@ -1,13 +1,9 @@
 from dataclasses import asdict
-from datetime import datetime
 from pathlib import PurePosixPath
 
 import pytest
-from ulid import ULID
 
 from renku_data_services.notebooks.api.classes.image import Image
-from renku_data_services.notebooks.core_sessions import get_mount_work_dir
-from renku_data_services.session.models import Environment, EnvironmentImageSource, EnvironmentKind, Member
 
 
 @pytest.mark.parametrize(
@@ -234,96 +230,3 @@ async def test_image_workdir_check(image: str, expected_path: PurePosixPath | No
         assert workdir is None, f"The image workdir should be None but instead it is {workdir}"
     else:
         assert workdir == expected_path
-
-
-def create_test_environment(
-    container_image: str, work_dir: PurePosixPath | None = None, mount_dir: PurePosixPath | None = None
-) -> Environment:
-    return Environment(
-        name="test",
-        container_image=container_image,
-        default_url="test",
-        id=ULID(),
-        environment_image_source=EnvironmentImageSource.image,
-        port=8888,
-        uid=1000,
-        gid=1000,
-        environment_kind=EnvironmentKind.CUSTOM,
-        creation_date=datetime.now(),
-        created_by=Member(id="id"),
-        build_parameters=None,
-        build_parameters_id=None,
-        mount_directory=mount_dir,
-        working_directory=work_dir,
-    )
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    "environment,expected_mount_dir,expected_work_dir",
-    [
-        (
-            create_test_environment("busybox:1.38"),
-            PurePosixPath("/work"),
-            PurePosixPath("/work"),
-        ),
-        (
-            create_test_environment("ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3"),
-            # The default work dir in the buildpack images is /workspace so we mount there
-            PurePosixPath("/workspace"),
-            PurePosixPath("/workspace"),
-        ),
-        (
-            # If we set only the work dir on the environment that should be used
-            create_test_environment(
-                "ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3",
-                work_dir=PurePosixPath("/home/renku/work"),
-            ),
-            PurePosixPath("/home/renku/work"),
-            PurePosixPath("/home/renku/work"),
-        ),
-        (
-            # If we set the work and mount dir on the environment that should be used
-            create_test_environment(
-                "ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3",
-                mount_dir=PurePosixPath("/home/renku/work"),
-                work_dir=PurePosixPath("/home/renku/work"),
-            ),
-            PurePosixPath("/home/renku/work"),
-            PurePosixPath("/home/renku/work"),
-        ),
-        (
-            # If the work dir is not inside the mount dir, the work dir will be set to the mount dir
-            create_test_environment(
-                "ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3",
-                mount_dir=PurePosixPath("/home/renku/mount"),
-                work_dir=PurePosixPath("/home/renku/work"),
-            ),
-            PurePosixPath("/home/renku/mount"),
-            PurePosixPath("/home/renku/mount"),
-        ),
-        (
-            create_test_environment(
-                "ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3",
-                mount_dir=PurePosixPath("/home/renku/mount"),
-            ),
-            PurePosixPath("/home/renku/mount"),
-            PurePosixPath("/home/renku/mount"),
-        ),
-        (
-            # If you try to mount on / we move it to /work to preven users from wiping out the whole image
-            create_test_environment(
-                "ghcr.io/swissdatasciencecenter/renku/py-datascience-jupyterlab:2.17.3",
-                mount_dir=PurePosixPath("/"),
-            ),
-            PurePosixPath("/work"),
-            PurePosixPath("/work"),
-        ),
-    ],
-)
-async def test_mount_workdir(
-    environment: Environment, expected_mount_dir: PurePosixPath, expected_work_dir: PurePosixPath
-) -> None:
-    mount_dir, work_dir = await get_mount_work_dir(environment)
-    assert mount_dir == expected_mount_dir
-    assert work_dir == expected_work_dir
